boost: code movement only: split CECKey into separate files
