+++ src/Makefile.am
@@ -82,6 +82,7 @@ BITCOIN_CORE_H = \
   core_io.h \
   crypter.h \
   db.h \
+  ecwrapper.h \
   hash.h \
   init.h \
   key.h \
@@ -211,6 +212,7 @@ libbitcoin_common_a_SOURCES = \
   core.cpp \
   core_read.cpp \
   core_write.cpp \
+  ecwrapper.cpp \
   hash.cpp \
   key.cpp \
   keystore.cpp \
